Add files via upload
MQ2 Gas Sensor: This gas sensor is made of a sensitive material called Tin Oxide also known as stannic oxide with the formula SnO2, this material has a low sensitivity in clean air, but when the target gas is present the sensor’s conductivity is higher and the gas concentration rising. This sensor has a high sensitivity to LPG, Propane and Hydrogen and can be used with Methane and other steam, see sensitivity characteristic of MQ2 of various gases in Fig.3 the graph below.  There are conditions which need to be avoided when using the sensor, for other Technical information see MQ2 Data sheet PDF format. Now to calculate the Pert per million(ppm) of the gases, this required getting two point the X and Y axis from fig.3, here are the formula to get the slope and the Y- intercept.

•	y = mx + b
•	Y: x value 
•	X: X Value
•	m: Slope of the line
•	b: Y intercept 
•	Using log-log scale the formula look like this:
•	Log(y) = m* log(x) + b
•	Let’s choose two points from the graph as to find the slope from the LPG point (200,1.8) and (10000, 0.28). The formula to calculate m is as followed:
•	m = [log(y) – log(y0)] / [log[x) – log(x0)]
•	Now let’s apply the logarithmic quotient rule as follow:
•	m = log(y/y0) /log(x/x0)
•	So, substitute the values for x, x0, y, and y0:
•	m = log(0.28/1.8) / log(10000/200)
•	m = -0.476
•	Next to calculate y intercept m is already know. So this step required choosing a point from the graph which is from the LPG line and the point is (5000,1.1)
•	Log(y) = m* log(x) + b 
•	b = log(y) – m* log(x)
•	b = log(1.1) – ( 0.476) * log(500)
•	b = 3.054
•	Now m and b are known, to find the gas concentration for any ratio see the formula:
•	log(x) = [log(y) – b] / m
•	Next to get the real value of the gas concentration according to the log-log plot the inverse of log x is needed:
•	X = 10^ {[log(y) – b] / m}. See more about logarithmic rule on this link: http://www.rapidtables.com/math/algebra/Logarithm.htm