Add files via upload
The project discuss in detail material like the MQ2 and MQ135 gas sensors, this section highlight MQ135 gas sensor as followed: 
MQ135 Gas Sensor: This gas sensor is composed of Aluminium Oxide (AL2O3) ceramic tube, Tin Dioxide (SnO2) sensitive layer, the measuring electrode and heater are fixed into a crust made by plastic and stainless-steel net. The resistance level is different to various kind of gas, when using the sensor, it is recommended that the required heat (leave on charge over 24hours) is provided in other to get accurate value of the results. The sensor has a high sensitivity to Ammonia with 100ppm resistance in clean air, Alcohol need the value of load resistance to be use for calibration, Benzene, smoke and CO2, see sensitivity characteristic of MQ135 various gases in Fig.4 the graph below. For other details see data sheet PDF format. Again, in other to calculate the R0 (Sensitivity Resistance) of the gases, use above formulas as the MQ2.